Reworked project controllers / implementing Command Pattern

+++ SWMS_3/controler/warehousing_controller.py
@@ -1,3 +1,5 @@
+from tkinter import messagebox
+
 from model.entities.warehouse import Warehouse
 from model.service.logger import MyLogger
 from model.service.modules.warehousing_module import WarehousingModule
@@ -8,7 +10,8 @@ class WarehousingController:
     def __init__(self, warehousing_module: WarehousingModule, logger: MyLogger):
         self.module = warehousing_module
         self.logger = logger
-        self.view = None
+        self.management_view = None
+        self.warehousing_view = None
 
     @property
     def warehouses(self):
@@ -49,7 +52,7 @@ class WarehousingController:
         result = self.module.create_wh(name, type_, capacity, [], status)
         if isinstance(result, Warehouse):
             self.reload()
-            self.view.refresh()
+            self.management_view.refresh()
         return result
 
     def update_warehouse(self, name, type_, capacity, products, status, id_):
@@ -57,25 +60,43 @@ class WarehousingController:
         result = self.module.update_warehouse(old, name, type_, capacity, products, status)
         if isinstance(result, Warehouse):
             self.reload()
-            self.view.refresh()
+            self.management_view.refresh()
         return result
 
-    def del_warehouse(self, selected):
+    def del_warehouse(self):
+        selected = self.management_view.item_list.get_selected_items()
+        if len(selected) < 1:
+            messagebox.showwarning("Warning!", "Please make a selection first!", parent=self.management_view.parent)
+            return
+
         wh_id = int(selected[0][0])
         result = self.module.delete_wh_by_id(wh_id)
         if isinstance(result, Warehouse):
             self.reload()
-            self.view.refresh()
-        return result
+            self.management_view.refresh()
+            messagebox.showinfo("Info!", f"Warehouse {result.name} successfully deleted!",
+                                parent=self.management_view.parent)
+        else:
+            messagebox.showwarning("Warning", result, parent=self.management_view.parent)
 
     # endregion
 
     # region GUI
-    def show_add_warehouse(self):
-        form = ItemForm(self.view.parent, Warehouse("", "", 0, [], ""), self, height=250)
+    def show_create_warehouse(self):
+        form = ItemForm(self.management_view.parent, Warehouse("", "", 0, [], ""), self, height=250)
 
-    def show_edit_warehouse(self, selected):
+    def show_edit_warehouse(self):
+        selected = self.management_view.item_list.get_selected_items()
+        if len(selected) < 1:
+            messagebox.showwarning("Warning!", "Please make a selection first!", parent=self.management_view.parent)
+            return
+
         wh_id = int(selected[0][0])
         warehouse = self.module.find_wh_by_id(wh_id)
-        form = ItemForm(self.view.parent, warehouse, self, height=250, edit=True)
+        form = ItemForm(self.management_view.parent, warehouse, self, height=250, edit=True)
+
     # endregion
+
+    def close(self):
+        self.management_view.open_views.remove(self.management_view.page_name)
+        self.management_view.parent.destroy()
